Make JUnitTest shut down its browser more reliably
