+++ .gitignore
@@ -4,6 +4,11 @@ out/
 # Transform jars built by Gradle
 data/qc/*.jar
 
+# Logs and temp files created by tests
+data/**/*.log
+data/**/assaydata/
+data/**/drt_temp/
+
 # Generated by test runner
 remainingTests.txt
